Moved Unsigned NTLM authentication as a core component to an example.
As a stub section for an  unsupported protocol that is Microsoft specific this would be better served as a stub example that we can build up if we encounter it again. These kind of specific examples will act both as training  guidance for creating report sections for new vulnerabilities and as existing language that we can pull from if auditors encounter these concerns in the wild.