Added old physical section which forced a  modified project structure.
The old physical section points out that a core component of physical security is password strength and encryption. It only makes sense to group these topics into the physical security section and leave the malware/network security of indiviudal devices in the local section. Audit organization and created content are reflected here.