Activity: Microsoft, escilation of privlage game added.
This game is a great security and vulnerability awareness activity. It allows creative exploration of possible vulnerabilities on a network. This gives us a way to make exploration of possible vulnerabilities a much more enjoyable endevor that prepares the auditors for the dryer exaustive research focus of the vulnerabiity research section of the training.