Removed passive OS detection from network scanning.
Pssive OS detection requires arpspoofing and therefore can lead to unintentional impacts on the organizations ability to do their job. Until we identify a way to do that does not impact a wide variety of organizations, like only spoofing one target, we will have to keep this out of the curricula.