Unlinked physical assesment from all index's.
The physical assessment component is deeply integrated with the user device assessment and data assessment. We have also found that it works just as well, if better, when done intermittantly as the auditor talks to auditees and walks around the building. For these reasons, and because of the heightened focus on digital security, we have chosen to remove physical assesments as a section in SAFETAG.