added reference
I'm not convinced that this variant should be separated from the forensic distribution variant ( evidence-capture/en/exercises/evidence_capture/variant_offline_using_forensic_distro.md ), as the usage of a forensic distro is the method of choice for dead imaging. I would create a single variant on dead imaging with this file and the file for capture with a forensic distro. Anyway, this commit is to add a reference to the other variant, if we decide to keep them separated.